NX: Continue with 64bit safety in a paranoid way

+++ source/platform/nx/common.c
@@ -157,11 +157,11 @@ int q_snprintf (char *str, size_t size, const char *format, ...)
 	return ret;
 }
 
-void Q_memset (void *dest, int fill, int count)
+void Q_memset (void *dest, int fill, size_t count)
 {
-	int             i;
+	size_t             i;
 	
-	if ( (((long)dest | count) & 3) == 0)
+	if ( (((size_t)dest | count) & 3) == 0)
 	{
 		count >>= 2;
 		fill = fill | (fill<<8) | (fill<<16) | (fill<<24);
@@ -173,11 +173,11 @@ void Q_memset (void *dest, int fill, int count)
 			((byte *)dest)[i] = fill;
 }
 
-void Q_memcpy (void *dest, void *src, int count)
+void Q_memcpy (void *dest, void *src, size_t count)
 {
-	int             i;
+	size_t             i;
 	
-	if (( ( (long)dest | (long)src | count) & 3) == 0 )
+	if (( ( (size_t)dest | (size_t)src | count) & 3) == 0 )
 	{
 		count>>=2;
 		for (i=0 ; i<count ; i++)
@@ -188,7 +188,7 @@ void Q_memcpy (void *dest, void *src, int count)
 			((byte *)dest)[i] = ((byte *)src)[i];
 }
 
-int Q_memcmp (void *m1, void *m2, int count)
+int Q_memcmp (void *m1, void *m2, size_t count)
 {
 	while(count)
 	{
